svc/api_gateway_service/requirements.txt
oauthlib==2.1.0
Flask-OAuthlib==0.9.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
mock==2.0.0
